refactor: modernize settings window
