+++ Harbor/Views/SettingsView.swift
@@ -5,214 +5,39 @@ struct SettingsView: View {
     @ObservedObject var updater: AppUpdater
 
     var body: some View {
-        @Bindable var settings = settings
-
-        Form {
-            Section("General") {
-                Toggle("Start downloads immediately", isOn: $settings.startDownloadsAutomatically)
-                Toggle("Send download notifications", isOn: $settings.notificationsEnabled)
-            }
-
-            Section("Destinations") {
-                DestinationFolderRow(
-                    title: "Regular Downloads",
-                    path: settings.defaultDestinationPath,
-                    choose: settings.chooseDefaultDestination,
-                    reveal: settings.revealDefaultDestination
-                )
-
-                DestinationFolderRow(
-                    title: "Torrent Downloads",
-                    path: settings.torrentDestinationPath,
-                    choose: settings.chooseTorrentDestination,
-                    reveal: settings.revealTorrentDestination
-                )
-            }
-
-            Section("Torrent Automation") {
-                Toggle("Watch a folder for torrent files", isOn: $settings.torrentWatchFolderEnabled)
-
-                DestinationFolderRow(
-                    title: "Watch Folder",
-                    path: settings.torrentWatchFolderPath,
-                    choose: settings.chooseTorrentWatchFolder,
-                    reveal: settings.revealTorrentWatchFolder
-                )
-
-                if settings.torrentWatchFolderEnabled {
-                    TorrentWatchStatusRow(status: settings.torrentWatchFolderStatus)
+        TabView {
+            GeneralSettingsTab(settings: settings)
+                .tabItem {
+                    Label("General", systemImage: "gearshape")
                 }
 
-                Toggle("Seed new torrents after downloading", isOn: $settings.seedNewTorrents)
-            }
-
-            Section("Bandwidth") {
-                Stepper(
-                    value: $settings.maxConcurrentDownloads,
-                    in: AppSettingsStore.maxConcurrentDownloadsRange
-                ) {
-                    LabeledContent("Max Active Downloads", value: "\(settings.maxConcurrentDownloads)")
+            DownloadsSettingsTab(settings: settings)
+                .tabItem {
+                    Label("Downloads", systemImage: "folder")
                 }
 
-                Stepper(
-                    value: $settings.perDownloadConnectionCount,
-                    in: AppSettingsStore.perDownloadConnectionCountRange
-                ) {
-                    LabeledContent("Connections per Download", value: "\(settings.perDownloadConnectionCount)")
+            TorrentsSettingsTab(settings: settings)
+                .tabItem {
+                    Label("Torrents", systemImage: "arrow.up.arrow.down.circle")
                 }
 
-                SpeedLimitRow(
-                    title: "Global Download Limit",
-                    isEnabled: $settings.globalSpeedLimitEnabled,
-                    kilobytesPerSecond: $settings.globalSpeedLimitKilobytesPerSecond
-                )
-
-                SpeedLimitRow(
-                    title: "Default Download Limit",
-                    isEnabled: $settings.perDownloadSpeedLimitEnabled,
-                    kilobytesPerSecond: $settings.perDownloadSpeedLimitKilobytesPerSecond
-                )
-
-                SpeedLimitRow(
-                    title: "Global Upload Limit",
-                    isEnabled: $settings.globalUploadSpeedLimitEnabled,
-                    kilobytesPerSecond: $settings.globalUploadSpeedLimitKilobytesPerSecond
-                )
-
-                SpeedLimitRow(
-                    title: "Default Torrent Upload Limit",
-                    isEnabled: $settings.perDownloadUploadSpeedLimitEnabled,
-                    kilobytesPerSecond: $settings.perDownloadUploadSpeedLimitKilobytesPerSecond
-                )
-            }
-
-            Section("Updates") {
-                LabeledContent("Current Version") {
-                    Text(updater.currentVersionLabel)
-                        .foregroundStyle(.secondary)
+            BandwidthSettingsTab(settings: settings)
+                .tabItem {
+                    Label("Bandwidth", systemImage: "speedometer")
                 }
 
-                Toggle(
-                    "Automatically check for updates",
-                    isOn: Binding(
-                        get: { updater.automaticallyChecksForUpdates },
-                        set: { updater.setAutomaticallyChecksForUpdates($0) }
-                    )
-                )
-
-                Button("Check for Updates…") {
-                    updater.checkForUpdates()
+            UpdatesSettingsTab(updater: updater)
+                .tabItem {
+                    Label("Updates", systemImage: "arrow.triangle.2.circlepath")
                 }
-                .buttonStyle(.borderedProminent)
-                .disabled(updater.canCheckForUpdates == false)
-            }
-        }
-        .formStyle(.grouped)
-    }
-}
 
-private struct DestinationFolderRow: View {
-    let title: LocalizedStringResource
-    let path: String
-    let choose: () -> Void
-    let reveal: () -> Void
-
-    var body: some View {
-        LabeledContent {
-            VStack(alignment: .trailing, spacing: 6) {
-                Text(path)
-                    .foregroundStyle(.secondary)
-                    .multilineTextAlignment(.trailing)
-                    .textSelection(.enabled)
-
-                HStack(spacing: 8) {
-                    Button("Choose…", action: choose)
-                    Button("Reveal", action: reveal)
+            AcknowledgmentsSettingsTab()
+                .tabItem {
+                    Label("Acknowledgments", systemImage: "info.circle")
                 }
-            }
-        } label: {
-            Text(title)
         }
-    }
-}
-
-private struct TorrentWatchStatusRow: View {
-    let status: TorrentWatchFolderStatus
-
-    var body: some View {
-        LabeledContent("Status") {
-            Label(message, systemImage: symbolName)
-                .foregroundStyle(foregroundStyle)
-        }
-    }
-
-    private var message: LocalizedStringResource {
-        switch status {
-        case .stopped:
-            "Waiting for the watcher to start"
-        case .watching:
-            "Watching for new torrent files"
-        case .unavailable:
-            "Folder unavailable; Harbor will retry"
-        }
-    }
-
-    private var symbolName: String {
-        switch status {
-        case .stopped:
-            "clock"
-        case .watching:
-            "eye"
-        case .unavailable:
-            "exclamationmark.triangle"
-        }
-    }
-
-    private var foregroundStyle: Color {
-        switch status {
-        case .stopped:
-            .secondary
-        case .watching:
-            .green
-        case .unavailable:
-            .orange
-        }
-    }
-}
-
-private struct SpeedLimitRow: View {
-    let title: LocalizedStringResource
-    @Binding var isEnabled: Bool
-    @Binding var kilobytesPerSecond: Int
-
-    var body: some View {
-        LabeledContent {
-            HStack(spacing: 8) {
-                Toggle("Limit", isOn: $isEnabled)
-                    .labelsHidden()
-                
-                TextField(
-                    "Speed",
-                    value: $kilobytesPerSecond,
-                    format: .number
-                )
-                .monospacedDigit()
-                .frame(width: 110)
-                .disabled(isEnabled == false)
-                
-                Text("KB/s")
-                    .foregroundStyle(.secondary)
-            }
-        } label: {
-            Text(title)
-                .lineLimit(1)
-        }
-        .onChange(of: kilobytesPerSecond) { _, newValue in
-            let clampedValue = AppSettingsStore.clampedSpeedLimitKilobytes(newValue)
-            if clampedValue != newValue {
-                kilobytesPerSecond = clampedValue
-            }
-        }
+        .frame(width: 640, height: 460)
+        .scenePadding()
     }
 }
 
@@ -221,6 +46,4 @@ private struct SpeedLimitRow: View {
         settings: HarborPreviewFixtures.makeSettings(),
         updater: AppUpdater.preview()
     )
-        .frame(width: 560, height: 760)
-        .padding(20)
 }
